Convert to having a stateful client
The commit introduces a large refactoring. Rather than having global
state storing the subscription information, there is now a type named
azure.Client, which is initialized with an azure.Config containing the
profile settings. Operations are then invoked on this client, meaning
that multiple subscriptions can be managed in one process.

In addition, the package names have been rationalized, and it is
necessary to pass the generic azure.Client to each of the specific
service clients.