Fix table entities error checking.
Some functions were only checking the response
code if a call to execTable returned an error,
but in some cases calls to this can fail
without returning errors. It is unclear if this
behaviour is intentiona or not, so for now we
just check the response code at all times.