V9 Generated (#569)
* Running latest Go Generator on all released packages.

* Gen.go update (#547)

* Updated Gen.go to use new autorest builld system

* Comments from review

* It works as expected again

* Moving Storage to an Independent Repository (#556)

* Updating CI to use Go 1.8.

Go 1.8 was released on Thursday, February 16th. Many of our partners
have adopted and we need to ensure that we support the most recent
version.

* Replacing storage folder with submodule.

In order to allow for more independent growth of the storage package,
we've given it its own repository. This change updates the main
repository to use a submodule in order to prevent breaking changes for
folks using storage today who do not want to migrate to the new
repository.

* Updating submodule target URL.

There is an existing naming convention for Azure Storage repositories.
the repo has been updated to match that convention, this updates the
submodule to point to the new URL.

* Responding to review feedback.

* Updating submodule to ref v8.1.0 storage version.

* Fixing issues found while reviewing PR.

* Regenerating using latest swaggers.

Using Azure/azure-rest-api-specs commit: 7d9b6c018d9de2b7de189ea8658963ef8d3bc41b

* Regenerating with new API Version strategy & web composite.

The new API Version strategy associated the API Version to send across the wire with each method instead of at the client level. Doing so allows us to support a broader range of composite swaggers (the immediate example is the web composite swagger.)

It also means that making a call with an alternate API version is somewhat more difficult in this version. To do so, one would have to create their own Preparer for the request. In the future, we'll support multiple API Versions and allow for more stability by generating independent packages for each Operation Group in each API Version. We expect to have that done in the summer.

* Regenerating against most recent swaggers.

* Regenerating using latest preview version of compute.