Add breaking changes detection to travis CI. (#1408)
Run the apidiff tool in travis to report breaking changes in a PR.  The
tool is run as an allowed failure and will return a non-zero exit code
if there are breaking changes so the task shows up as red.  The raw JSON
log will also be output to the console.