Refactor azcore (#15349)
* Refactor azcore

See the CHANGELOG for details.

* add additional test coverage

move early header check back to Pipeline.Do

* add nil body check

* added more test coverage

removed spurious error return from NewPoller() sig.
fixed bug in FinalResponse() to perform the final GET if the URL is
specified but the terminal response contains no content.

* fix tests that don't return a final payload

* remove bad examples

* move constants to make version info easier to find

* moved Poller constructors out of root packages

for core, moved to azcore/runtime
for arm, moved to azcore/arm/runtime
also moved ARM RP registration policy to runtime

In order to do this, it was necessary to move all pipeline related
content to its own internal package azcore/internal/pipeline.  This was
to avoid problems with circular dependencies.  It also groups things
that are logically related and cleaned up the code.