[azservicebus] Changing ReceivedMessage.Body() to ReceivedMessage.Body (a field) (#17888)
Changed ReceivedMessage.Body() to be a field of type `[]byte`. 

The entire reason it was a function was to return an error in the rare case that the underlying AMQP message was using the Value or Sequence section, rather than Data (or that Data itself was multiple byte arrays, not just a single one). Being a function makes it harder for users to create tests where they pass around `ReceivedMessage` since they can't create it themselves.

In the end it was agreed that just having a nil Body was enough signal for users to know that something was wrong, and that the error that was returned wouldn't have really been actionable (much like Body being nil, it was basically a true/false with exactly one error message).

Thanks to @ItalyPaleAle for pointing this out!

Fixes #17877